+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 python-asserts adheres to [semantic versioning](https://semver.org/).
 
-## UNRELEASED –
+## 0.14.0 – 2025-11-17
 
 ### Added